* create and mount postgres connect helper * substitute repeated code for postgres connection using helper
